refactor: Move the checkpoint list tests to their own file
